+++ springline/src/settings/governance_settings.rs
@@ -16,10 +16,10 @@ pub struct GovernanceSettings {
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct GovernanceRuleSettings {
-    pub min_cluster_size: u16,
-    pub max_cluster_size: u16,
-    pub min_scaling_step: u16,
-    pub max_scaling_step: u16,
+    pub min_cluster_size: u32,
+    pub max_cluster_size: u32,
+    pub min_scaling_step: u32,
+    pub max_scaling_step: u32,
     pub custom: telemetry::TableType,
 }
